LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/performance-timeline.https-expected.txt



Harness Error (TIMEOUT), message = null

PASS Test Performance Timeline API in Service Worker 
FAIL empty service worker fetch event included in performance timings assert_greater_than: Slow service worker request should measure increased delay. expected a number greater than 1061 but got 61
PASS User Timing 
TIMEOUT Resource Timing Test timed out
